refactor: UserProfilerCard를 컴포넌트로 추출
