+++ components/molecules/index.ts
@@ -9,3 +9,4 @@ export { default as ExhibitionInfo } from './ExhibitionInfo';
 export { default as ReviewExhibitionInfo } from './ReviewExhibitionInfo';
 export { default as ScrollButton } from './ScrollButton';
 export { default as SideNavigation } from './SideNavigation';
+export { default as UserProfileCard } from './UserProfileCard';
